fix: reorganizando a pasta de models e removendo flags desnecessárias na conexão com o banco
