+++ src/database/index.js
@@ -8,10 +8,7 @@ class Database {
   init() {
     try {
       mongoose.connect(process.env.MONGO_URL, {
-        useCreateIndex: true,
         useNewUrlParser: true,
-        useFindAndModify: true,
-        useUnifiedTopology: true,
       });
       console.log('Database connected!');
     } catch (e) {
